Refactor MediaOverlay models into separate files
Split MediaOverlay-related classes into individual files with clear namespaces for better organization and maintainability. Move utility classes to EpubReader.Util. Update EpubReader.csproj to use MSBuild properties for Android signing credentials, removing hardcoded secrets. Remove local NuGet source from nuget.config and clean up unused using directives. These changes enhance code clarity, security, and portability.
