+++ EpubReader/Models/MediaOverlayAudio.cs
@@ -1,0 +1,3 @@
+namespace EpubReader.Models.MediaOverlays;
+
+public sealed record MediaOverlayAudio(string Source, TimeSpan? ClipBegin, TimeSpan? ClipEnd);
