+++ EpubReader/Models/MediaOverlayParallel.cs
@@ -1,0 +1,8 @@
+namespace EpubReader.Models.MediaOverlays;
+
+public sealed class MediaOverlayParallel : MediaOverlayNode
+{
+	public MediaOverlayText? Text { get; init; }
+
+	public MediaOverlayAudio? Audio { get; init; }
+}
