+++ EpubReader/Models/MediaOverlaySequence.cs
@@ -1,0 +1,8 @@
+namespace EpubReader.Models.MediaOverlays;
+
+public sealed class MediaOverlaySequence : MediaOverlayNode
+{
+	public string? TextReference { get; init; }
+
+	public List<MediaOverlayNode> Children { get; } = [];
+}
