+++ EpubReader/Models/MediaOverlayText.cs
@@ -1,0 +1,3 @@
+namespace EpubReader.Models.MediaOverlays;
+
+public sealed record MediaOverlayText(string Source);
