+++ EpubReader/Service/MediaOverlayPlaybackManager.cs
@@ -1,6 +1,5 @@
 using System.Text.Encodings.Web;
 using System.Text.Json;
-using EpubReader.Models;
 using EpubReader.Models.MediaOverlays;
 using Plugin.Maui.Audio;
