+++ nuget.config
@@ -4,7 +4,6 @@
 		<!--To inherit the global NuGet package sources remove the <clear/> line below -->
 		<clear />
 		<add key="nuget" value="https://api.nuget.org/v3/index.json" />
-		<add key="Local" value="C:\Users\james\Downloads\Nugets" />
 		<add key="CommunityToolkitGitHubNuget" value="https://nuget.pkg.github.com/CommunityToolkit/index.json" />
 	</packageSources>
 	<!-- CommunityToolkit GitHub Packages feed is added conditionally in CI
